invoice-processing-pipeline/reviewer/requirements.txt
flask>=2.0.0
google-auth>=2.0.0
google-cloud-firestore>=2.0.0
google-cloud-storage>=2.0.0
gunicorn>=20.0.0